Define default method for on-fail
Note that there seems to be a bug where, even though all the tests are
working as expected, the output of the on-fail method (called as part
of a test) is being printed to the REPL. This shouldn't happen because
a with-output-to-string form rebinds *standard-output*, but the output
is both being sent to the string AND being printed to the
original *standard-output*.

The cause seems to be the second test, as deleting it makes the bug
disappear. But I have no idea why its presence would cause this
behavior.